refactor(grimoire): isolate navigator state
